+++ .github/CHANGELOG.md
@@ -41,6 +41,9 @@
 
 ### Breaking changes
 
+* Update `lightning.gpu` and `lightning.kokkos` to raise an error instead of falling back to `default.qubit`.
+  [(#689)](https://github.com/PennyLaneAI/pennylane-lightning/pull/689)
+
 * Add `paths` directives to test workflows to avoid running tests that cannot be impacted by changes.
   [(#699)](https://github.com/PennyLaneAI/pennylane-lightning/pull/699)
   [(#695)](https://github.com/PennyLaneAI/pennylane-lightning/pull/695)